#1514 Moved GOceanLoopSwapTrans into its own file under domain/...
