+++ doc/user_guide/gocean1p0.rst
@@ -909,7 +909,7 @@ the GOcean 1.0 API. For an overview of transformations in general see
 
 ####
 
-.. autoclass:: psyclone.transformations.GOLoopSwapTrans
+.. autoclass:: psyclone.domain.gocean.transformations.GOLoopSwapTrans
    :members:
    :noindex:
