+++ src/psyclone/domain/gocean/transformations/__init__.py
@@ -48,14 +48,16 @@ from psyclone.domain.gocean.transformations.gocean_loop_fuse_trans \
     import GOceanLoopFuseTrans
 from psyclone.domain.gocean.transformations.gocean_const_loop_bounds_trans \
     import GOConstLoopBoundsTrans
+from psyclone.domain.gocean.transformations.go_loop_swap_trans \
+    import GOLoopSwapTrans
 
 # The entities in the __all__ list are made available to import directly from
 # this package e.g.:
 # from psyclone.domain.gocean.transformations import GOceanExtractTrans
 
 __all__ = ['GOceanExtractTrans',
-           'GOMoveIterationBoundariesInsideKernelTrans',
            'GOceanLoopFuseTrans',
-           'GOOpenCLTrans',
            'GOConstLoopBoundsTrans',
-           'GOceanLoopFuseTrans']
+           'GOLoopSwapTrans',
+           'GOMoveIterationBoundariesInsideKernelTrans',
+           'GOOpenCLTrans']
