+++ src/psyclone/transformations.py
@@ -2208,132 +2208,6 @@ class Dynamo0p3RedundantComputationTrans(LoopTrans):
         loop.update_halo_exchanges()
 
 
-class GOLoopSwapTrans(LoopTrans):
-    ''' Provides a loop-swap transformation, e.g.:
-
-    .. code-block:: fortran
-
-        DO j=1, m
-            DO i=1, n
-
-    becomes:
-
-    .. code-block:: fortran
-
-        DO i=1, n
-            DO j=1, m
-
-    This transform is used as follows:
-
-     >>> from psyclone.parse.algorithm import parse
-     >>> from psyclone.psyGen import PSyFactory
-     >>> ast, invokeInfo = parse("shallow_alg.f90")
-     >>> psy = PSyFactory("gocean1.0").create(invokeInfo)
-     >>> schedule = psy.invokes.get('invoke_0').schedule
-     >>> # Uncomment the following line to see a text view of the schedule
-     >>> # schedule.view()
-     >>>
-     >>> from psyclone.transformations import GOLoopSwapTrans
-     >>> swap = GOLoopSwapTrans()
-     >>> swap.apply(schedule.children[0])
-     >>> # Uncomment the following line to see a text view of the schedule
-     >>> # schedule.view()
-
-    '''
-    def __str__(self):
-        return "Exchange the order of two nested loops: inner becomes " + \
-               "outer and vice versa"
-
-    def validate(self, node_outer, options=None):
-        '''Checks if the given node contains a valid Fortran structure
-        to allow swapping loops. This means the node must represent
-        a loop, and it must have exactly one child that is also a loop.
-
-        :param node_outer: a Loop node from an AST.
-        :type node_outer: py:class:`psyclone.psyir.nodes.Loop`
-        :param options: a dictionary with options for transformations.
-        :type options: dict of string:values or None
-
-        :raises TransformationError: if the supplied node does not\
-                                     allow a loop swap to be done.
-
-        '''
-        super(GOLoopSwapTrans, self).validate(node_outer, options=options)
-
-        if not isinstance(node_outer, GOLoop):
-            raise TransformationError("Error in GOLoopSwap transformation. "
-                                      "Given node '{0}' is not a GOLoop, but "
-                                      "an instance of '{1}."
-                                      .format(node_outer, type(node_outer)))
-
-        if not node_outer.loop_body or not node_outer.loop_body.children:
-            raise TransformationError("Error in GOLoopSwap transformation. "
-                                      "Supplied node '{0}' must be the outer "
-                                      "loop of a loop nest and must have one "
-                                      "inner loop, but this node does not "
-                                      "have any statements inside."
-                                      .format(node_outer))
-
-        node_inner = node_outer.loop_body[0]
-
-        # Check that the body of the outer loop is itself a Loop
-        try:
-            super(GOLoopSwapTrans, self).validate(node_inner, options=options)
-        except TransformationError as err:
-            six.raise_from(
-                TransformationError("Error in GOLoopSwap transformation. "
-                                    "Supplied node '{0}' must be the outer "
-                                    "loop of a loop nest but the first "
-                                    "inner statement is not a valid loop:\n"
-                                    "{1}.".format(node_outer, str(err.value))),
-                err)
-
-        if not isinstance(node_inner, GOLoop):
-            raise TransformationError(
-                "Error in GOLoopSwap transformation. Inner loop of supplied "
-                "loop nest ({0}) is not a GOLoop, but an instance of '{1}'."
-                .format(node_outer, type(node_inner).__name__))
-
-        if len(node_outer.loop_body.children) > 1:
-            raise TransformationError(
-                "Error in GOLoopSwap transformation. Supplied node '{0}' must"
-                " be the outer loop of a loop nest and must have exactly one "
-                "inner loop, but this node has {1} inner statements, the "
-                "first two being '{2}' and '{3}'"
-                "".format(node_outer, len(node_outer.loop_body.children),
-                          node_outer.loop_body[0], node_outer.loop_body[1]))
-
-    def apply(self, outer, options=None):
-        # pylint: disable=arguments-differ
-        '''The argument :py:obj:`outer` must be a loop which has exactly
-        one inner loop. This transform then swaps the outer and inner loop.
-
-        :param outer: the node representing the outer loop.
-        :type outer: :py:class:`psyclone.psyir.nodes.Loop`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if the supplied node does not \
-                                     allow a loop swap to be done.
-
-        '''
-        self.validate(outer, options=options)
-
-        schedule = outer.root
-        inner = outer.loop_body[0]
-
-        # Detach the inner code
-        inner_loop_body = inner.loop_body.detach()
-
-        # Swap the loops
-        outer.replace_with(inner.detach())
-        inner.addchild(Schedule())
-        inner.loop_body.addchild(outer)
-
-        # Insert again the inner code in the new inner loop
-        outer.loop_body.replace_with(inner_loop_body)
-
-
 class Dynamo0p3AsyncHaloExchangeTrans(Transformation):
     '''Splits a synchronous halo exchange into a halo exchange start and
     halo exchange end. For example:
@@ -3364,7 +3238,6 @@ __all__ = ["KernelTrans",
            "ACCParallelTrans",
            "MoveTrans",
            "Dynamo0p3RedundantComputationTrans",
-           "GOLoopSwapTrans",
            "Dynamo0p3AsyncHaloExchangeTrans",
            "Dynamo0p3KernelConstTrans",
            "ACCEnterDataTrans",
